Extract settings account, sync, and reset orchestration hooks
